Shrink image size, detect window, and show faces
We don't *really* need 640x480 images, do we?  Let's scale it
back to 320x240.  That means we have to scale back our detection
windows from 50x50 to something smaller.  Done.

While we're at it, let's show individual windows with only the face
rectangles of each detection.  Kinda creepy, but useful.

Lastly, the default haar rectangles are extreme-close-up territory.
Let's move the detected window up a few pixels and expand it down
further so we can capture an entire face and not just eyes, nose,
and a closed mouth.